% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{calculate_included_consensus}
\alias{calculate_included_consensus}
\title{calculate_included_consensus}
\usage{
calculate_included_consensus(review_results_df, review_info)
}
\arguments{
\item{review_results_df}{the dataframe with the review data from get_review_results_df() API call and tidying function
rename_included_cols}
}
\value{
the R object containing the dataframe with a new column 'customizations_included_consensus' with the consensus where there is one (or 'Conflict' if not)
}
\description{
calculates consensus based on individual reviewer decisions
}
\keyword{internal}
