% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inclusion_counts.R
\name{facets}
\alias{facets}
\title{get_reviews_raw}
\usage{
facets(api_tokens, id)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}
}
\value{
the R object containing the result of the API call
}
\description{
gets reviews from the rayyan API and outputs an R object
}
\keyword{internal}
