% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{get_aws_presigned_url}
\alias{get_aws_presigned_url}
\title{get_aws_presigned_url}
\usage{
get_aws_presigned_url(api_tokens, review_id)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}

\item{review_id}{the rayyan ID of the review to get - this can be obtained via
get_reviews}
}
\value{
the R object containing the result of the API call
}
\description{
gets a URL to enable uploading of files or artilces
}
\keyword{internal}
