% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inclusion_counts.R, R/get_review_info_raw.R
\name{get_review_info_raw}
\alias{get_review_info_raw}
\title{get_review_info_raw}
\usage{
get_review_info_raw(api_tokens, review_id)

get_review_info_raw(api_tokens, review_id)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}

\item{review_id}{the rayyan ID of the review to get - this can be obtained via
get_reviews}

\item{id}{the rayyan ID of the review to get - this can be obtained via
get_reviews}
}
\value{
the R object containing the result of the API call

the R object containing the result of the API call
}
\description{
gets a review from the rayyan API and outputs the metadata an R object

gets a review from the rayyan API and outputs the metadata an R object
}
\keyword{internal}
