% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_review_results_raw.R
\name{get_review_results_raw}
\alias{get_review_results_raw}
\title{get_review_results_raw}
\usage{
get_review_results_raw(api_tokens, review_id)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}

\item{review_id}{the rayyan ID of the review to get - this can be obtained via
get_reviews}
}
\value{
the R object containing the result of the API call
}
\description{
gets a review from the rayyan API and outputs the results an R object
}
\keyword{internal}
