% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inclusion_counts.R, R/reviews.R
\name{get_reviews}
\alias{get_reviews}
\title{get_reviews}
\usage{
get_reviews(api_tokens)

get_reviews(api_tokens)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}
}
\value{
a dataframe containing the ID, title, and owner of reviews

a dataframe containing the ID, title, and owner of reviews
}
\description{
gets the reviews from the rayyan API and returns the ID, title,
and owner in a dataframe

gets the reviews from the rayyan API and returns the ID, title,
and owner in a dataframe
}
