% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_user_info}
\alias{get_user_info}
\title{get_user_info}
\usage{
get_user_info(api_env)
}
\arguments{
\item{api_env}{the api envrionment from load_tokens_and_env()}
}
\value{
an object containing the user information from the rayyan API.
}
\description{
Gets the user information from rayyan
}
