% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{list_reviews_raw}
\alias{list_reviews_raw}
\title{list_reviews_raw}
\usage{
list_reviews_raw(api_env)
}
\arguments{
\item{api_env}{the api envrionment from load_tokens_and_env()
or login_tokens_and_env()}
}
\value{
the R object containing the result of the API call
}
\description{
gets reviews from the rayyan API and outputs an R object
}
\keyword{internal}
