% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tokens.R
\name{load_tokens}
\alias{load_tokens}
\title{load_tokens}
\usage{
load_tokens(creds)
}
\arguments{
\item{creds}{local path to the json file containing the credentials exported from rayyan}
}
\value{
an api environment to be passed to the other functions
}
\description{
loads API credentials from a JSON file
}
