% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tokens.R
\name{load_tokens_and_env}
\alias{load_tokens_and_env}
\title{load_tokens_and_env}
\usage{
load_tokens_and_env(creds, base_url = "https://rayyan.ai")
}
\arguments{
\item{creds}{the json file containing the credentials exported from rayyan}

\item{base_url}{the API base URL (defaults to production)}
}
\value{
an api environment to be passed to the other functions
}
\description{
loads API credentials from a JSON file
and returns an object containing the tokens and
API URL
}
