% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_tokens.R
\name{login_tokens}
\alias{login_tokens}
\title{login_tokens}
\usage{
login_tokens(access_token, refresh_token)
}
\arguments{
\item{creds}{local path to the json file containing the credentials exported from rayyan}
}
\value{
an api environment to be passed to the other functions
}
\description{
imports the API tokens from function arguments
}
