% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tokens.R
\name{login_tokens_and_env}
\alias{login_tokens_and_env}
\title{login_tokens_and_env}
\usage{
login_tokens_and_env(
  access_token,
  refresh_token,
  base_url = "https://rayyan.ai"
)
}
\arguments{
\item{base_url}{the API base URL (defaults to production)}

\item{creds}{the json file containing the credentials exported from rayyan}
}
\value{
an api environment to be passed to the other functions
}
\description{
imports the API tokens from function arguments
and returns an object containing the tokens and
API URL
}
