% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tokens.R
\name{refresh_tokens}
\alias{refresh_tokens}
\title{refresh_tokens}
\usage{
refresh_tokens(creds, update_local = T)
}
\arguments{
\item{creds}{local path to the json file containing the credentials exported from rayyan}

\item{update_local}{option to overwrite existing json file containing the credentials with the new, updated credentials. Defaults to TRUE.}
}
\value{
an api environment to be passed to the other functions
}
\description{
refreshes the API tokens from function arguments
}
