% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_included_cols_names.R
\name{rename_included_cols_names}
\alias{rename_included_cols_names}
\title{rename_included_cols_names}
\usage{
rename_included_cols_names(
  review_results_df,
  review_info,
  rename_with = "name"
)
}
\arguments{
\item{review_results_df}{the dataframe with the review data from get_review_results_df() API call and tidying function}

\item{review_info}{a list with the review data from get_review_info() API call function}

\item{rename_with}{what should the IDs be renamed with. Must be 'name' or 'email'.
rename_included_cols}
}
\value{
the R object containing the tidied dataframe
}
\description{
renames the included columns' names so that they have explicit reviewer names rather than reviewer id
}
\keyword{internal}
