% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_included_cols_values.R
\name{rename_included_cols_values}
\alias{rename_included_cols_values}
\title{rename_included_cols_values}
\usage{
rename_included_cols_values(review_results_df)
}
\arguments{
\item{review_results_df}{the dataframe with the review data from get_review_results_df() API call and tidying function
rename_included_cols}
}
\value{
the R object containing the tidied dataframe
}
\description{
renames the included columns' values so they are explicitly 'Excluded' (-1), 'Maybe' (0), and 'Included' (1)
}
\keyword{internal}
