% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewresults.R
\name{reviewresults}
\alias{reviewresults}
\title{reviewresults}
\usage{
reviewresults(
  api_tokens,
  review_id = NULL,
  stages = c("both", "records"),
  recordsreview_id = NULL,
  reportsreview_id = NULL
)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env() or login_tokens_and_env()}

\item{review_id}{the rayyan ID of the review to get - this can be obtained via}

\item{stages}{specify if both stages present in the review}

\item{recordsreview_id}{if there is a separate review for the records/abstract (rather than it being in), specify it here'.}

\item{reportsreview_id}{if there is a separate review for the reports/full text (rather than it being in), specify it here'.}
}
\value{
the R object containing the result of the API call #'
}
\description{
gets a review from the rayyan API and outputs the results a tidied R object
}
\keyword{internal}
