% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewresults_calculateconsensus.R
\name{reviewresults_calculateconsensus}
\alias{reviewresults_calculateconsensus}
\title{reviewresults_calculateconsensus}
\usage{
reviewresults_calculateconsensus(review_results_df, review_info)
}
\arguments{
\item{review_results_df}{the dataframe with the review data from get_review_results_df() API call and tidying function
rename_included_cols}
}
\value{
the R object containing the dataframe with a new column 'customizations_included_consensus' with the consensus where there is one (or 'conflict' if not)
}
\description{
calculates consensus based on individual reviewer decisions
}
\keyword{internal}
