% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_all.R
\name{unnest_all}
\alias{unnest_all}
\title{unnest_all}
\usage{
unnest_all(df)
}
\arguments{
\item{df}{the dataframe that came from the json file}
}
\value{
the R object containing the unnested dataframe
}
\description{
recursively unnests all the list columns in a dataframe. borrowed from:
https://stackoverflow.com/questions/63786411/how-to-unnest-wider-with-loop-over-all-the-columns-containing-lists
}
\keyword{internal}
