% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{get_review_info_raw}
\alias{get_review_info_raw}
\title{get_review_info_raw}
\usage{
get_review_info_raw(api_env, id)
}
\arguments{
\item{api_env}{the api environment from load_tokens_and_env()
or login_tokens_and_env()}

\item{id}{the rayyan ID of the review to get - this can be obtained via
get_reviews}
}
\value{
the R object containing the result of the API call
}
\description{
gets a review from the rayyan API and outputs the metadata an R object
}
\keyword{internal}
