% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayyan2PRISMA2020.R
\name{rayyan2PRISMA2020}
\alias{rayyan2PRISMA2020}
\title{rayyan2PRISMA2020}
\usage{
rayyan2PRISMA2020(
  identification_databases = NULL,
  identification_duplicates = NULL,
  screening_recordsandreports = NULL,
  nofulltext_string = "No full text available",
  exclusionreasons_reports = NULL,
  included_studies_n = NULL
)
}
\arguments{
\item{identification_databases}{a list of search databases}

\item{identification_duplicates}{database of duplicates}

\item{screening_recordsandreports}{database of review results for all screened records, with results at both records and reports stage for each record}

\item{nofulltext_string}{string used for denoting no full text available}

\item{exclusionreasons_reports}{vector of strings for exclusion reasons to be tallied}

\item{included_studies_n}{number of included studies (not reports) at end}
}
\value{
the R object containing the result of the API call
}
\description{
gets a review from the rayyan API and outputs the results a tidied R object
}
\keyword{internal}
