% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewresults.R
\name{reviewresults}
\alias{reviewresults}
\title{reviewresults}
\usage{
reviewresults(
  api_tokens,
  recordsandreports_review_id = NULL,
  records_review_id = NULL,
  reports_review_id = NULL
)
}
\arguments{
\item{api_tokens}{the api environment from load_tokens_and_env() or login_tokens_and_env()}

\item{recordsandreports_review_id}{f there is a single review containing both the records/abstract and reports/full text stages (intended implementation of Rayyan), specify it here (N.B. This is currently a placeholder and won't work as there is currently no API endpoint from which to read the full text stage of a Rayyan review, as far as I can tell)}

\item{records_review_id}{if there is a separate review for the records/abstract (rather than it being in a single review with the reports/full text stage), specify it here'.}

\item{reports_review_id}{if there is a separate review for the reports/full text (rather than it being in a single review with the records/abstract stage), specify it here'.}

\item{review_id}{the rayyan ID of the review to get - this can be obtained via}
}
\value{
the R object containing the result of the API call #'
}
\description{
gets a review from the rayyan API and outputs the results a tidied R object
}
\keyword{internal}
